231004 <update note> 1. 밀리모드, 레인지모드 스킬 8개 구성 2. 액션바 쿨타임기능 3. 스킬 소리 삽입

+++ Assets/CODE/Enemy/Enemys.cs
@@ -106,7 +106,7 @@ public class Enemys : MonoBehaviour
     bool Enemy_Hit;
     bool KB;
     public float KBSpeed;
-    public void F_OnHIt(int _DMG)
+    public void F_OnHIt(float _DMG)
     {
         if (CurHP > 0 && !Enemy_Hit)
         {
